Sediment Source Analysis

Sediment Source Analysis

Functions

Index

Functions

SedimentAnalysis.MTF.combined_norm — Function

combined_norm(u, v, ...)

Compute the combined norm of the arguments as if all arguments were part of one large array.

This is equivilent to norm(cat(u, v, ...)), but this implimentation avoids creating anintermediate array.

u = [3 0]
v = [0 4 0]
combined_norm(u, v)

# output

5.0

SedimentAnalysis.MTF.nnmtf — Function

nnmtf(Y::Abstract3Tensor, R::Integer; kwargs...)

Non-negatively matrix-tensor factorizes an order 3 tensor Y with a given "rank" R.

Factorizes $Y \approx C F$ where $\displaystyle Y[i,j,k] \approx \sum_{r=1}^R C[i,r]*F[r,j,k]$ and the factors $C, F \geq 0$ are nonnegative.

Note there may NOT be a unique optimal solution

Arguments

Y::Abstract3Tensor: tensor to factorize

R::Integer: rank to factorize Y (size(C)[2] and size(F)[1])

Keywords

maxiter::Integer=100: maxmimum number of iterations

tol::Real=1e-3: desiered tolerance for the -gradient's distance to the normal cone

rescale::Bool=false: scale F at each iteration so that the 3-fiber sums are 1 (on average)

plot_F::Integer=0: if not 0, plot F every plot_F iterations

names::AbstractVector{String}=String[]: names of the slices of F to use for ploting

Returns

C::Matrix{Float64}: the matrix C in the factorization Y ≈ C * F

F::Array{Float64, 3}: the tensor F in the factorization Y ≈ C * F

rel_errors::Vector{Float64}: relative errors at each iteration

norm_grad::Vector{Float64}: norm of the full gradient at each iteration

dist_Ncone::Vector{Float64}: distance of the -gradient to the normal cone at each iteration

SedimentAnalysis.MTF.plot_factors — Function

plot_factors(F, names; appendtitle="")

Plot each horizontal slice of F. Names give the name of each vertical slice.

SedimentAnalysis.MTF.rel_error — Function

rel_error(x, xhat)

Compute the relative error between x (true value) and xhat (its approximation).

The relative error is given by:

\[\frac{\lVert x - \hat{x} \rVert}{\lVert x \rVert}\]

Index

SedimentAnalysis.MTF.combined_norm

SedimentAnalysis.MTF.nnmtf

SedimentAnalysis.MTF.plot_factors

SedimentAnalysis.MTF.rel_error