Matrix Tensor Factorization

Types

SedimentAnalysis.MTF.Abstract3Tensor — Type

Alias for an AbstractArray{T, 3}.

Functions

Base.:* — Method

Base.*(A::AbstractMatrix, B::Abstract3Tensor)

Computes the Abstract3Tensor C where $C_{ijk} = \sum_{l=1}^L A_{il} * B_{ljk}$.

When the third dimention of B has length 1, this is equivilent to the usual matrix-matrix multiplication. For this reason, we resuse the same symbol.

This is equivilent to the $1$-mode product $B \times_1 A$.

SedimentAnalysis.MTF.combined_norm — Function

combined_norm(u, v, ...)

Compute the combined norm of the arguments as if all arguments were part of one large array.

This is equivilent to norm(cat(u, v, ...)), but this implimentation avoids creating an intermediate array.

u = [3 0]
v = [0 4 0]
combined_norm(u, v)

# output

5.0

SedimentAnalysis.MTF.dist_to_Ncone — Function

dist_to_Ncone(grad_C, grad_F, C, F)

Calculate the distance of the -gradient to the normal cone of the positive orthant.

SedimentAnalysis.MTF.nnmtf — Function

nnmtf(Y::Abstract3Tensor, R::Integer; kwargs...)

Non-negatively matrix-tensor factorizes an order 3 tensor Y with a given "rank" R.

Factorizes $Y \approx C F$ where $\displaystyle Y[i,j,k] \approx \sum_{r=1}^R C[i,r]*F[r,j,k]$ and the factors $C, F \geq 0$ are nonnegative.

Note there may NOT be a unique optimal solution

Arguments

Y::Abstract3Tensor: tensor to factorize

R::Integer: rank to factorize Y (size(C)[2] and size(F)[1])

Keywords

maxiter::Integer=100: maxmimum number of iterations

tol::Real=1e-3: desiered tolerance for the -gradient's distance to the normal cone

rescale_CF::Bool=true: scale F at each iteration so that the factors (horizontal slices) have similar 3-fiber sums.

rescale_Y::Bool=true: Preprocesses the input Y to have normalized 3-fiber sums (on average), and rescales the final F so Y=C*F.

plot_F::Integer=0: if not 0, plot F every plot_F iterations

names::AbstractVector{String}=String[]: names of the slices of F to use for ploting

Returns

C::Matrix{Float64}: the matrix C in the factorization Y ≈ C * F

F::Array{Float64, 3}: the tensor F in the factorization Y ≈ C * F

rel_errors::Vector{Float64}: relative errors at each iteration

norm_grad::Vector{Float64}: norm of the full gradient at each iteration

dist_Ncone::Vector{Float64}: distance of the -gradient to the normal cone at each iteration

SedimentAnalysis.MTF.plot_factors — Function

plot_factors(F, names; appendtitle="")

Plot each horizontal slice of F. Names give the name of each vertical slice.

SedimentAnalysis.MTF.rel_error — Function

rel_error(x, xhat)

Compute the relative error between x (true value) and xhat (its approximation).

The relative error is given by:

\[\frac{\lVert \hat{x} - x \rVert}{\lVert x \rVert}\]

See also mean_rel_error.

SedimentAnalysis.MTF.mean_rel_error — Function

mean_rel_error(X, Xhat; dims=(1,2))

Compute the mean relative error between the dims-order slices of X and Xhat.

The mean relative error is given by:

\[\frac{1}{N}\sum_{j=1}^N\frac{\lVert \hat{X}_j - X_j \rVert}{\lVert X_j \rVert}\]

See also rel_error.

Approximations

SedimentAnalysis.MTF.d_dx — Function

d_dx(y::AbstractVector{<:Real})

Approximates the 1nd derivative of a function using only given samples y of that function.

Assumes y came from f(x) where x was an evenly sampled, unit intervel grid. Note the approximation uses centered three point finite differences for the next-to-end points, and foward/backward three point differences for the begining/end points respectively. The remaining interior points use five point differences.

Force only the third order approximation with thirdorder=true. See [`d2dx2`](@ref).

SedimentAnalysis.MTF.d2_dx2 — Function

d2_dx2(y::AbstractVector{<:Real})

Approximates the 2nd derivative of a function using only given samples y of that function.

Assumes y came from f(x) where x was an evenly sampled, unit intervel grid. Note the approximation uses centered three point finite differences for the next-to-end points, and foward/backward three point differences for the begining/end points respectively. The remaining interior points use five point differences.

Force only the third order approximation with thirdorder=true. See [`ddx`](@ref).

SedimentAnalysis.MTF.curvature — Function

curvature(y::AbstractVector{<:Real})

Approximates the signed curvature of a function given evenly spaced samples.

Uses d_dx and d2_dx2 to approximate the first two derivatives.

SedimentAnalysis.MTF.standard_curvature — Function

standard_curvature(y::AbstractVector{<:Real})

Approximates the signed curvature of a function, scaled to the unit box $[0,1]^2$.

See curvature.

Index

SedimentAnalysis.SedimentTools.DEFAULT_ALPHA

SedimentAnalysis.SedimentTools.DEFAULT_N_SAMPLES

SedimentAnalysis.MTF.Abstract3Tensor

SedimentAnalysis.SedimentTools.DensityTensor

SedimentAnalysis.SedimentTools.Grain

SedimentAnalysis.SedimentTools.Rock

SedimentAnalysis.SedimentTools.Sink

SedimentAnalysis.SedimentTools.Source

Base.:*

Base.getindex

Base.names

Plots.heatmap

SedimentAnalysis.MTF.combined_norm

SedimentAnalysis.MTF.curvature

SedimentAnalysis.MTF.d2_dx2

SedimentAnalysis.MTF.d_dx

SedimentAnalysis.MTF.dist_to_Ncone

SedimentAnalysis.MTF.mean_rel_error

SedimentAnalysis.MTF.nnmtf

SedimentAnalysis.MTF.plot_factors

SedimentAnalysis.MTF.rel_error

SedimentAnalysis.MTF.standard_curvature

SedimentAnalysis.SedimentTools.array

SedimentAnalysis.SedimentTools.default_bandwidth

SedimentAnalysis.SedimentTools.eachdensity

SedimentAnalysis.SedimentTools.eachmeasurement

SedimentAnalysis.SedimentTools.eachsink

SedimentAnalysis.SedimentTools.eachsource

SedimentAnalysis.SedimentTools.estimate_which_source

SedimentAnalysis.SedimentTools.getdomain

SedimentAnalysis.SedimentTools.getdomains

SedimentAnalysis.SedimentTools.getmeasurements

SedimentAnalysis.SedimentTools.getsink

SedimentAnalysis.SedimentTools.getsource

SedimentAnalysis.SedimentTools.getsourcename

SedimentAnalysis.SedimentTools.getsourcenames

SedimentAnalysis.SedimentTools.getstepsizes

SedimentAnalysis.SedimentTools.label_accuracy

SedimentAnalysis.SedimentTools.make_densities

SedimentAnalysis.SedimentTools.match_sources!

SedimentAnalysis.SedimentTools.measurement_heatmaps

SedimentAnalysis.SedimentTools.namedarray

SedimentAnalysis.SedimentTools.normalize_density_sums

SedimentAnalysis.SedimentTools.normalize_density_sums!

SedimentAnalysis.SedimentTools.plot_convergence

SedimentAnalysis.SedimentTools.plot_densities

SedimentAnalysis.SedimentTools.plot_source_index

SedimentAnalysis.SedimentTools.read_raw_data

SedimentAnalysis.SedimentTools.setsourcename!

SedimentAnalysis.SedimentTools.source_heatmaps

SedimentAnalysis.SedimentTools.standardize_KDEs