Sediment Analysis Tools

Types

SedimentAnalysis.SedimentTools.Grain — Type

Grain(v::AbstractVector{T}, measurement_names::AbstractVector{String})

Struct to hold grain level data

SedimentAnalysis.SedimentTools.DensityTensor — Type

DensityTensor(KDEs, domains, sinks)
DensityTensor(array, domains, measurement_names; kw...)

An order 3 array to hold the density distributions for multiple sinks.

KDEs is a Vector{Vector{Vector{T}}} like type whereas array is an Array{T,3} like type.

Call setsourcename! to set the source name (name of first dimention).

SedimentAnalysis.SedimentTools.Rock — Type

Alias for Sink

SedimentAnalysis.SedimentTools.Sink — Type

Sink(grain1, grain2, ...)
Sink([grain1, grain2, ...])

Collects a list of Grains into a Rock/Sink.

Ensures all Grains have the same names and are in the same order. Struct to hold sink level data

SedimentAnalysis.SedimentTools.Source — Type

Alias for Sink

Constants

SedimentAnalysis.SedimentTools.DEFAULT_ALPHA — Constant

Smoothing parameter for calculating a kernel's bandwidth

SedimentAnalysis.SedimentTools.DEFAULT_N_SAMPLES — Constant

Number of samples to use when standardizing a vector of density estimates.

Method Extentions

Base.getindex — Method

Gets all values of the measurement k in the Sink.

Base.names — Method

names(n::NamedArray, dimname::Union{String,Symbol})

Extend the names function from NamedArray to get the names given the axis name rather than the axis number.

Plots.heatmap — Method

Plots.heatmap(M::NamedMatrix; kwargs...)

Automatically grabs the names in axis to label the x and y plot axis

Functions

Importers

SedimentAnalysis.SedimentTools.read_raw_data — Function

read_raw_data(filename)
read_raw_data(filename; skip_sheets)

Imports excel data to a Vector{Sink}.

Excel file must have one element per page where different columns correspond to different sinks. Each sink can have a different number of grains (length of the column), but a sink must have a consistant length across different measurements (sheets).

Optionaly provide a collection skip_sheets to blacklist sheet names from the excel file.

Getters

SedimentAnalysis.SedimentTools.array — Function

Turn a DensityTensor or NamedArray into a plain Array type.

SedimentAnalysis.SedimentTools.getdomain — Function

getdomain(D::DensityTensor, measurement::String)
getdomain(D::DensityTensor, j::Integer)

Gets the domain for the measurement density, the locations where the density was sampled. See getdomains.

SedimentAnalysis.SedimentTools.getdomains — Function

getdomains(D::DensityTensor{T})::Vector{Vector{T}}

Gets the domain for every measurement's density, the locations where each density was sampled. See getdomains.

SedimentAnalysis.SedimentTools.getmeasurements — Function

getmeasurements(g::Grain)
getmeasurements(s::Sink)
getmeasurements(D::DensityTensor)

Getter for the measurement names.

Gets the names of measurements from a Sink

SedimentAnalysis.SedimentTools.getsourcename — Function

getsourcename(D::DensityTensor)

Gets the name for the grouping of measurements. Usually "Sink" or "Source". See getsourcename.

SedimentAnalysis.SedimentTools.getsourcenames — Function

getsourcenames(D::DensityTensor)

Gets the list of all sources' names. For example, ["Sink 1", "Sink 2", ...]. See getsourcenames.

SedimentAnalysis.SedimentTools.getstepsizes — Function

getstepsizes(D::DensityTensor)

Gets the step sizes used for each domain. See getdomains.

SedimentAnalysis.SedimentTools.namedarray — Function

Turn a DensityTensor into a NamedArray type.

SedimentAnalysis.SedimentTools.getsink — Function

Alias for getsource

SedimentAnalysis.SedimentTools.getsource — Function

getsource(D::DensityTensor, i::Integer)
getsink(D::DensityTensor, i::Integer)

Gets source/sink i from D. See eachsource.

Setters and Manipulators

SedimentAnalysis.SedimentTools.normalize_density_sums! — Function

normalize_density_sums!(D::DensityTensor)

Rescales the densities so that the sum of the density samples is 1.

This is in constrast to the usualy normalization for density functions where the area of the density curve is 1. In the case of an evenly sampled density, this area is sum(density_samples)*step_size.

Use normalize_density_sums to avoid mutation.

SedimentAnalysis.SedimentTools.normalize_density_sums — Function

See normalize_density_sums!

SedimentAnalysis.SedimentTools.setsourcename! — Function

setsourcename!(D::DensityTensor, name::String)

Sets the name of the source used by getsourcename and getsourcenames.

Iterators

SedimentAnalysis.SedimentTools.eachdensity — Function

eachdensity(D::DensityTensor)
eachdensity(D::DensityTensor, measurement::String)

Iterates D over each density vector. These are the 3 fibers of D. If a measurement is given, iterates over the densities for that measurement.

SedimentAnalysis.SedimentTools.eachmeasurement — Function

Iterator for a list of values of each measurement

eachmeasurement(D::DensityTensor)

Iterates D over each measurement slice. These are the lateral slices.

SedimentAnalysis.SedimentTools.eachsink — Function

Alias for eachsource.

SedimentAnalysis.SedimentTools.eachsource — Function

eachsource(D::DensityTensor)
eachsink(D::DensityTensor)

Iterates D over each source/sink slice. These are the horizontal slices. See getsource.

Density and Source Estimation

SedimentAnalysis.SedimentTools.default_bandwidth — Function

default_bandwidth(data; alpha=0.9, inner_percentile=100)

Coppied from KernelDensity since this function is not exported. I want access to it so that the same bandwidth can be used for different densities for the same measurements.

SedimentAnalysis.SedimentTools.estimate_which_source — Function

estimate_which_source(grain::Grain, F::DensityTensor; kwargs...)

Returns the likelihood and source index of the mostly likely factor the grain vector came from.

Returns

(Default) source_index::Integer: The index of the most likely source

(when max_likelihoods==true) (maxlikelihood, source_index): The most likely source and its likelihood

(when all_likelihoods==true) likelihoods::Vector{Real}: Likelihood grain came from each source

(when both are true) ((maxlikelihood, source_index), likelihoods)

SedimentAnalysis.SedimentTools.label_accuracy — Function

label_accuracy(labels, true_amounts::AbstractArray{T,3})

Calculates the number of correctly labeled grains and percentage of correct labels

Arguments

labels: Iterable of iterables with the labels for each grain (Vector{Vector{T}} or Tuple{Tuple})

true_amounts: Number of grains from each source for every sink

Outputs

n_correct_eachsink::Integer

n_total_labels::Integer

accuracy::Real`

SedimentAnalysis.SedimentTools.make_densities — Function

make_densities(s::Sink; kwargs...)
make_densities(s::Sink, domains::AbstractVector{<:AbstractVector}; kwargs...)

Estimates the densities for each measurement in a Sink.

When given domains, a list where each entry is a domain for a different measurement, resample the kernel on this domain.

Parameters

bandwidths::AbstractVector{<:Real}: list of bandwidths used for each measurement's

density estimation

inner_percentile::Integer=100: value between 0 and 100 that filters out each measurement

by using the inner percentile range. This can help remove outliers and focus in on where the bulk of the data is.

Returns

density_estimates::Vector{UnivariateKDE}

SedimentAnalysis.SedimentTools.match_sources! — Function

match_sources!(C, F, C_true, F_true)

Permute sources in C and F to match the ground truth Ctrue and Ftrue.

Similarity is checked by finding the source that minimizes norm(c - c_true) where c is the column of C.

Parameters

double_check::Bool: When true, repeat for F and F_true with their horizontal slices, and assert the ordering is the same as before.

SedimentAnalysis.SedimentTools.standardize_KDEs — Function

standardize_KDEs(KDEs::AbstractVector{UnivariateKDE}; n_samples=DEFAULT_N_SAMPLES,)

Resample the densities so they all are smapled from the same domain.

Resample the densities within each sink/source so that like-measurements use the same scale.

Visualizers

SedimentAnalysis.SedimentTools.measurement_heatmaps — Function

measurement_heatmaps(D::DensityTensor; kw...)

Returns heatmaps for each measurement (lateral slices) of D.

SedimentAnalysis.SedimentTools.plot_densities — Function

distributions_plot(D::DensityTensor, measurement::String; kw...)

Returns one plot will all distributions for a given measurement (lateral slice) of D.

SedimentAnalysis.SedimentTools.source_heatmaps — Function

source_heatmaps(D::DensityTensor; title="", kw...)

Returns heatmaps for each source (horizontal slices) of D.

The source name/index for each source will be appended to title.

SedimentAnalysis.SedimentTools.plot_convergence — Function

plot_convergence(rel_errors, norm_grad, dist_Ncone)

Returns 3 separate plots for the three convergence metrics on a log10-y scale.

plot_convergence((rel_errors, norm_grad, dist_Ncone))

SedimentAnalysis.SedimentTools.plot_source_index — Function

plot_source_index(
indexes::AbstractVector{<:Integer},
loglikelihood_ratios::AbstractVector{<:Real};
kwargs...
)

Returns one scatter plot of dots (eachindex(indexes), indexes) with brighter colours corresponding to higher loglikelihood_ratios.

Index

SedimentAnalysis.SedimentTools.DEFAULT_ALPHA

SedimentAnalysis.SedimentTools.DEFAULT_N_SAMPLES

SedimentAnalysis.MTF.Abstract3Tensor

SedimentAnalysis.SedimentTools.DensityTensor

SedimentAnalysis.SedimentTools.Grain

SedimentAnalysis.SedimentTools.Rock

SedimentAnalysis.SedimentTools.Sink

SedimentAnalysis.SedimentTools.Source

Base.:*

Base.getindex

Base.names

Plots.heatmap

SedimentAnalysis.MTF.combined_norm

SedimentAnalysis.MTF.curvature

SedimentAnalysis.MTF.d2_dx2

SedimentAnalysis.MTF.d_dx

SedimentAnalysis.MTF.dist_to_Ncone

SedimentAnalysis.MTF.mean_rel_error

SedimentAnalysis.MTF.nnmtf

SedimentAnalysis.MTF.plot_factors

SedimentAnalysis.MTF.rel_error

SedimentAnalysis.MTF.standard_curvature

SedimentAnalysis.SedimentTools.array

SedimentAnalysis.SedimentTools.default_bandwidth

SedimentAnalysis.SedimentTools.eachdensity

SedimentAnalysis.SedimentTools.eachmeasurement

SedimentAnalysis.SedimentTools.eachsink

SedimentAnalysis.SedimentTools.eachsource

SedimentAnalysis.SedimentTools.estimate_which_source

SedimentAnalysis.SedimentTools.getdomain

SedimentAnalysis.SedimentTools.getdomains

SedimentAnalysis.SedimentTools.getmeasurements

SedimentAnalysis.SedimentTools.getsink

SedimentAnalysis.SedimentTools.getsource

SedimentAnalysis.SedimentTools.getsourcename

SedimentAnalysis.SedimentTools.getsourcenames

SedimentAnalysis.SedimentTools.getstepsizes

SedimentAnalysis.SedimentTools.label_accuracy

SedimentAnalysis.SedimentTools.make_densities

SedimentAnalysis.SedimentTools.match_sources!

SedimentAnalysis.SedimentTools.measurement_heatmaps

SedimentAnalysis.SedimentTools.namedarray

SedimentAnalysis.SedimentTools.normalize_density_sums

SedimentAnalysis.SedimentTools.normalize_density_sums!

SedimentAnalysis.SedimentTools.plot_convergence

SedimentAnalysis.SedimentTools.plot_densities

SedimentAnalysis.SedimentTools.plot_source_index

SedimentAnalysis.SedimentTools.read_raw_data

SedimentAnalysis.SedimentTools.setsourcename!

SedimentAnalysis.SedimentTools.source_heatmaps

SedimentAnalysis.SedimentTools.standardize_KDEs